+++ src/views-components/dialog-forms/copy-collection-dialog.ts
@@ -6,7 +6,7 @@ import { compose } from "redux";
 import { withDialog } from "~/store/dialog/with-dialog";
 import { reduxForm } from 'redux-form';
 import { COLLECTION_COPY_FORM_NAME, CollectionCopyFormDialogData, copyCollection } from '~/store/collections/collection-copy-actions';
-import { DialogCollectionCopy } from "~/views-components/dialog-collection-copy/dialog-collection-copy";
+import { DialogCollectionCopy } from "~/views-components/dialog-copy/dialog-collection-copy";
 
 export const CopyCollectionDialog = compose(
     withDialog(COLLECTION_COPY_FORM_NAME),
